Resolve "edytowanie i usuwanie produktów do których użytkownik ma uprawnienia"

Closes #12

See merge request kp.marczynski/dietify!13

+++ src/main/java/pl/marczynski/dietify/products/service/impl/ProductServiceImpl.java
@@ -1,18 +1,22 @@
 package pl.marczynski.dietify.products.service.impl;
 
+import javassist.NotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pl.marczynski.dietify.core.domain.User;
+import pl.marczynski.dietify.core.service.UserService;
+import pl.marczynski.dietify.core.web.rest.errors.OperationNotAllowedForCurrentUserException;
 import pl.marczynski.dietify.products.domain.Product;
 import pl.marczynski.dietify.products.repository.ProductRepository;
 import pl.marczynski.dietify.products.service.ProductService;
 import pl.marczynski.dietify.products.service.ProductSubcategoryService;
 
-import java.util.Objects;
 import java.util.Optional;
 
 /**
@@ -27,11 +31,13 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final CacheManager cacheManager;
     private final ProductSubcategoryService productSubcategoryService;
+    private final UserService userService;
 
-    public ProductServiceImpl(ProductRepository productRepository, CacheManager cacheManager, ProductSubcategoryService productSubcategoryService) {
+    public ProductServiceImpl(ProductRepository productRepository, CacheManager cacheManager, ProductSubcategoryService productSubcategoryService, UserService userService) {
         this.productRepository = productRepository;
         this.cacheManager = cacheManager;
         this.productSubcategoryService = productSubcategoryService;
+        this.userService = userService;
     }
 
     /**
@@ -43,10 +49,25 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public Product save(Product product) {
         log.debug("Request to save Product : {}", product);
+        if (!hasRightsToPersistProduct(product)) {
+            throw new OperationNotAllowedForCurrentUserException();
+        }
         this.clearProductCaches(product);
+        if (product.getAuthor() == null) {
+            product.setAuthor(userService.getCurrentUser().get());
+        }
         Product result = productRepository.saveAndFlush(product);
         productSubcategoryService.removeOrphans();
         return result;
+    }
+
+    private boolean hasRightsToPersistProduct(Product product) {
+        Optional<User> currentUser = userService.getCurrentUser();
+        if (!currentUser.isPresent()) {
+            return false;
+        } else if (product.getId() == null || product.getAuthor() == null) {
+            return true;
+        } else return product.getAuthor().equals(currentUser.get());
     }
 
     /**
@@ -91,8 +112,15 @@ public class ProductServiceImpl implements ProductService {
      * @param id the id of the entity
      */
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws NotFoundException {
         log.debug("Request to delete Product : {}", id);
+        Optional<Product> product = productRepository.findOneWithEagerRelationships(id);
+        if (!product.isPresent()) {
+            throw new NotFoundException("Product not found");
+        }
+        if (!hasRightsToPersistProduct(product.get())) {
+            throw new OperationNotAllowedForCurrentUserException();
+        }
         this.clearProductCaches(id);
         productRepository.deleteById(id);
         productSubcategoryService.removeOrphans();
@@ -105,6 +133,9 @@ public class ProductServiceImpl implements ProductService {
     }
 
     private void clearProductCaches(long productId) {
-        Objects.requireNonNull(cacheManager.getCache(ProductRepository.PRODUCTS_EAGER_BY_ID_CACHE)).evict(productId);
+        Cache cache = cacheManager.getCache(ProductRepository.PRODUCTS_EAGER_BY_ID_CACHE);
+        if (cache != null) {
+            cache.evict(productId);
+        }
     }
 }
